turn on go-etcd logging
Latest changes to go-etcd cause the client
to cycle through endpoints and retry automatically.
That is great except that no logs are printed while
that is going on. This change will make flannel a lot
more chatty though. A move to etcd/client is planned
that should make logging more sane.