Automatically detects the Route Table ID
Route table ID is not required to be specified in the config. We
determine the subnetID the instance belongs to by using
describeInstances. We filter the DescribeRouteTables
request using the subnetID to find the RouteTableID. Autodetect functionality can be
overridden by manually passing routeTableID in as a configuration parameter.

Fixes #194