Add GCE backend
No configuration parameter is required other than the Backend=gce.
On startup, flannel deletes any conflicting entries from the route table and adds an entry setting the next hop, for the subnet it has acquired,
to itself. Instead of NextHopIP we use NextHopInstance. In the rare case
that a conflicting rule does exist, flannel waits for the write to be
implemented, which takes an average of 20 seconds, before inserting the
new entry.

Fixes #176