remote: ensure testing HTTP requests are canceled
Due to https://github.com/golang/go/issues/11013 race conditions
in the remote testcases mean that HTTP requests aren't actually
canceled.  In client.go::httpDo() the goroutine that actually
runs the http.Do() may not execute until after that routine's
select{} loop, at which point the Context may have already been
canceled.  Unfortunately with Go < 1.5 there is a race in the
http.Transport package that drops cancellations on the floor.

This issue can be reproduced by using the remote_test.go hunks
that use a WaitGroup to wait for subnet.WatchLeases() to exit
after the Context is canceled, but not the client.go hunks.
Specifically, doTestWatch() calls cancel() before
watch.go::WatchLeases() calls sm.WatchLeases(), which in this
case is RemoteManager.WatchLeases(), which then calls
RemoteManager.httpGet() -> httpDo(), which then calls
http.Client.Do() which has the aformentioned cancellation race.

To fix the races more completely (though not as completely
as Go 1.5 will do) we have to copy most of http.Transport
into the tree and implement some cancellation fixups ourselves.
A few parts of the new Transport are notable:

1) CancelRequest() - if the request is not yet being tracked,
the cancelation is saved in another map for later

2) getConn() - the real work is done here; when adding a
new request to the internal table, check if it the client has
asked to cancel it, and actually cancel it.  Second if we
are going to use an idle connection for the request, and the
request has already been canceled, then cancel the idle
connection too.

Second, use a WaitGroup in doTestWatch() to assert that the
outstanding watch is actually canceled.