Have registry deal with subnet and not etcd types
Let the registry expose the interface that uses types
like subnet.Lease instead of etcd.Response. This will
make it easier to support something other than etcd if
a need arises. And it makes the subnet manager simpler
as it has gotten to big lately.

Unrelated change: the initialization logic will retry
from the beginning on failer, not the step that failed.

Conflicts:
	subnet/local_manager.go