Bug fix: running out of memory with vxlan+bonding
Due to a kernel bug, vxlan backend would run out of
memory if there was bonding on the machine. There
would be infinite number of fdb entries returned when
queried. This works around this by filtering for
interface in kernel in addition to userspace.
This is better anyway.

Due to the urgency of the bug, the netlink commit
has not yet been upstreamed.

Fixes #367