Add functional (end-to-end) testing
This script is the first cut at getting end-to-end testing
for flannel. It brings up etcd in a container and then 2 flannel
instances (also in containers). It then pings one flannel instance
from another. It does this for each of the three backends being
tested: udp, vxlan, host-gw.

To run the script, use "make e2e-test"