dist/*: Add the compiled binaries for iptables and libpthread
These haven't changed since they were added and make the builds very
slow
Also, netfilter.org is down which breaks the iptables build.

All of the build code is still in place, so it should be simple to bump
the versions if we ever need to, or to build from source again.