Add iptables binaries
These haven't changed since they were added and rebuilding them makes the builds
very slow
Also, netfilter.org sometimes goes down which makes everything break.

All of the build code is still in place, so it should be simple to bump
the versions if we ever need to, or to build from source again.

(This adds about ~2MB to the compressed git repo)