subnet/*: Remove unused reservations code
The concept of a reservation was added a long time ago but never exposed
to users. This commit removes that code.

Reservations were part of the subnet manager interface but only
implemented by the etcd backend. The Kube backend had stub code to
implement the interface but threw errors when called.