edit Flannel license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit leaves the LICENSE file as-is and updates three other files
so that Licensee correctly recognizes the LICENSE file as the only
license-related file. The README's "License" section has been renamed
to "Licensing" and license-check.sh has been renamed to header-check.sh.
In the Makefile, the references to license-check.sh have been renamed
accordingly.

Collectively, these changes allow Licensee to successfully identify the
license type of Flannel as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>