docs: deploy flannel with NET_RAW
NET_RAW is a capability that docker has given containers by default forever. However, it's not needed for the majority of pod workloads, and is a security hole for workloads that don't need it.
CRI-O dropped NET_RAW by default in 1.18
I believe pods that need capabilities should ask for them, and we shouldn't let every pod have caps they don't need.
This PR adds NET_RAW to flannel's requested capabilities. For Docker, it will be a no-op; for CRI-O, the pod will successfully come up

Signed-off-by: Peter Hunt <pehunt@redhat.com>