ip: improve primary address selection to account for address flags
This change uses known IFA_F flags that are attached to interface
addresses by the linux kernel to differentiate between the usability
of multiple addresses.
Primarily, this deprioritizes temporary and deprecated addresses
that have been assigned to a node through RFC7217 address generation
techniques. It also explicitly prioritizes addresses that are marked
as "manually assigned" by the kernel.

This improves routing stability for all backends that require an
address to reach a node.

See #2293 for an example of why this is needed.

Signed-off-by: lilly <li@lly.sh>